requirements.txt
streamlit>=1.44.0
numpy>=1.26.4
scikit-learn>=1.5.0
pypdf>=5.4.0
python-docx>=1.1.2
pytest>=8.3.5

